Fully document the `wasmparser` crate.
This commit fully documents the `wasmparser` crate and denies missing
documentation.
