+++ crates/wasmparser/src/readers/data_section.rs
@@ -18,22 +18,32 @@ use super::{
     SectionReader, SectionWithLimitedItems,
 };
 
+/// Represents a data segment in a core WebAssembly module.
 #[derive(Debug, Copy, Clone)]
 pub struct Data<'a> {
+    /// The kind of data segment.
     pub kind: DataKind<'a>,
+    /// The data of the data segment.
     pub data: &'a [u8],
+    /// The range of the data segment.
     pub range: Range,
 }
 
+/// The kind of data segment.
 #[derive(Debug, Copy, Clone)]
 pub enum DataKind<'a> {
+    /// The data segment is passive.
     Passive,
+    /// The data segment is active.
     Active {
+        /// The memory index for the data segment.
         memory_index: u32,
+        /// The initialization expression for the data segment.
         init_expr: InitExpr<'a>,
     },
 }
 
+/// A reader for a core WebAssembly module's data section.
 #[derive(Clone)]
 pub struct DataSectionReader<'a> {
     reader: BinaryReader<'a>,
@@ -42,6 +52,7 @@ pub struct DataSectionReader<'a> {
 }
 
 impl<'a> DataSectionReader<'a> {
+    /// Constructs a new `DataSectionReader` for the given data and offset.
     pub fn new(data: &'a [u8], offset: usize) -> Result<DataSectionReader<'a>> {
         let mut reader = BinaryReader::new_with_offset(data, offset);
         let count = reader.read_var_u32()?;
@@ -52,14 +63,17 @@ impl<'a> DataSectionReader<'a> {
         })
     }
 
+    /// Gets the original position of the section reader.
     pub fn original_position(&self) -> usize {
         self.reader.original_position()
     }
 
+    /// Gets the count of items in the section.
     pub fn get_count(&self) -> u32 {
         self.count
     }
 
+    /// Whether or not to forbid data segments using bulk memory proposal.
     pub fn forbid_bulk_memory(&mut self, forbid: bool) {
         self.forbid_bulk_memory = forbid;
     }
