+++ crates/wasmparser/src/readers/export_section.rs
@@ -18,13 +18,18 @@ use super::{
     SectionWithLimitedItems,
 };
 
+/// Represents an export from a core WebAssembly module.
 #[derive(Debug, Copy, Clone)]
 pub struct Export<'a> {
+    /// The name of the exported item.
     pub field: &'a str,
+    /// The kind of the export.
     pub kind: ExternalKind,
+    /// The index of the exported item.
     pub index: u32,
 }
 
+/// A reader for a core WebAssembly module's export section.
 #[derive(Clone)]
 pub struct ExportSectionReader<'a> {
     reader: BinaryReader<'a>,
@@ -32,16 +37,19 @@ pub struct ExportSectionReader<'a> {
 }
 
 impl<'a> ExportSectionReader<'a> {
+    /// Constructs a new `ExportSectionReader` for the given data and offset.
     pub fn new(data: &'a [u8], offset: usize) -> Result<ExportSectionReader<'a>> {
         let mut reader = BinaryReader::new_with_offset(data, offset);
         let count = reader.read_var_u32()?;
         Ok(ExportSectionReader { reader, count })
     }
 
+    /// Gets the original position of the section reader.
     pub fn original_position(&self) -> usize {
         self.reader.original_position()
     }
 
+    /// Gets the count of items in the section.
     pub fn get_count(&self) -> u32 {
         self.count
     }
