+++ crates/wasmparser/src/readers/table_section.rs
@@ -18,6 +18,7 @@ use super::{
     TableType,
 };
 
+/// A reader for a core WebAssembly table section.
 #[derive(Clone)]
 pub struct TableSectionReader<'a> {
     reader: BinaryReader<'a>,
@@ -25,16 +26,19 @@ pub struct TableSectionReader<'a> {
 }
 
 impl<'a> TableSectionReader<'a> {
+    /// Constructs a new `TableSectionReader` for the given data and offset.
     pub fn new(data: &'a [u8], offset: usize) -> Result<TableSectionReader<'a>> {
         let mut reader = BinaryReader::new_with_offset(data, offset);
         let count = reader.read_var_u32()?;
         Ok(TableSectionReader { reader, count })
     }
 
+    /// Gets the original position of the section reader.
     pub fn original_position(&self) -> usize {
         self.reader.original_position()
     }
 
+    /// Gets the count of items in the section.
     pub fn get_count(&self) -> u32 {
         self.count
     }
